% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_to_factor.R
\name{char_to_factor}
\alias{char_to_factor}
\title{Converts all character variables to factor variables in a data frame}
\usage{
char_to_factor(df)
}
\arguments{
\item{df}{A data frame or tibble object.}
}
\value{
the same dataframe provided with all character variables converted to factor variables
}
\description{
When supplied a data frame with character variables, this function
converts all character variables within a data frame to factor variables.
}
