% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_low_levels.R
\name{convert_low_levels}
\alias{convert_low_levels}
\title{Convert Level with Low Frequencies to 'other'}
\usage{
convert_low_levels(x, thresh, new_level_name = "Other")
}
\arguments{
\item{x}{factor or character vector}

\item{thresh}{threshold for which all levels with frequency less than will be categorized into 'Other' or 'new_level_name'}

\item{new_level_name}{optional parameter to define what the new all encompasing level name should be}
}
\value{
A factor vector
}
\description{
This function takes a vector input 'x' that should be of type factor or character.
It converts all levels with frequency < 'thresh' to an "Other" category (by default),
or to a name of the users choosing specified by 'new_level_name'
}
