% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_binary.R
\name{is_binary}
\alias{is_binary}
\title{Logical Check if variable is binary or not}
\usage{
is_binary(y)
}
\arguments{
\item{y}{A numeric, integer, or character vector}
}
\value{
Logical TRUE or FALSE. If TRUE, the supplied vector is binary (contains only 2 levels when assessed as a factor)
}
\description{
This function takes a vector input 'y' and checks if it is binary
}
