% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_allsheets.R
\name{read_excel_allsheets}
\alias{read_excel_allsheets}
\title{Reads in all sheets from an Excel File and stores them as a list}
\usage{
read_excel_allsheets(filename, tibble = FALSE)
}
\arguments{
\item{filename}{the filename of the excel file to read in}

\item{tibble}{logical; if TRUE, then reads in as TIBBLE.}
}
\value{
a list of dataframes or TIBBLEs corresponding to each Excel Sheet
}
\description{
When supplied an excel file, this reads in all the sheets and stores each as a
component to a list.
}
